Fixed gcc 2 build.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@35981 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/system/kernel/debug/debug.cpp
@@ -1101,7 +1101,7 @@ cmd_dump_syslog(int argc, char** argv)
 		// remove '\0' and 0xcc
 		size_t toPrint = 0;
 		for (size_t i = 0; i < bytesRead; i++) {
-			if (buffer[i] != '\0' && buffer[i] != 0xcc)
+			if (buffer[i] != '\0' && (uint8)buffer[i] != 0xcc)
 				buffer[toPrint++] = buffer[i];
 		}
